+++ packages/vue/src/sdk.ts
@@ -1,8 +1,10 @@
 import { getDefaultIntegrations, init as browserInit } from '@sentry/browser';
 import type { Client } from '@sentry/core';
-import { applySdkMetadata } from '@sentry/core';
+import { applySdkMetadata, setNormalizeStringifier } from '@sentry/core/browser';
+
 import { vueIntegration } from './integration';
 import type { Options } from './types';
+import { normalizeStringifyValue } from './normalizeStringifyValue';
 
 /**
  * Inits the Vue SDK
@@ -15,5 +17,10 @@ export function init(options: Partial<Omit<Options, 'tracingOptions'>> = {}): Cl
 
   applySdkMetadata(opts, 'vue');
 
-  return browserInit(opts);
+  const client = browserInit(opts);
+
+  // Add vue-specific stringification
+  setNormalizeStringifier(normalizeStringifyValue);
+
+  return client;
 }